fix.7.5.2023 MiniProject - [Project Final Update]  - Content :    1. Controller      - ItemController : parameter 부분 image쪽 한칸 띄었음    2. Dto      - ItemDto : minPriceWanted Type change[Integer => Long]    3. Entity      - ItemEntity : minPriceWanted Type change[Integer => Long]    4. Service      - ItemService : Image 부분 프로필 폴더 생성 하여 이미지 저장해주는 경로로 다시 재설정함[예외발생도 포함]      - NegotiationService : 수정부분 text 수정

+++ market/src/main/java/com/likelion/market/dto/ItemDto.java
@@ -11,7 +11,7 @@ public class ItemDto {
     private String title;
     private String description;
     private String imageUrl;
-    private Integer minPriceWanted;
+    private Long minPriceWanted;
     private String status;
     private String writer;
     private String password;
