+++ market/src/main/java/com/likelion/market/entity/ItemEntity.java
@@ -15,7 +15,7 @@ public class ItemEntity {
     private String title;
     private String description;
     private String imageUrl;
-    private Integer minPriceWanted;
+    private Long minPriceWanted;
     private String status;
 
     private String writer;
